requirements.txt
annotated-types==0.7.0
anyio==4.9.0
audiooop-lts==0.2.1
blinker==1.9.0
certifi==2025.4.26
chardet==5.2.0
click==8.1.8
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.2.5
openai==1.77.0
pandas==2.2.3
pdfkit==1.0.0
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-magic==0.4.27
python-magic-bin==0.4.14; platform_system == "Windows"
pytz==2025.2
qrcode==8.2
regex==2024.11.6
reportlab==4.4.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.2
SQLAlchemy==2.0.40
standard-aifc==3.13.0
standard-chunk==3.13.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
